pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "openenv-codered"
version = "0.1.0"
description = "CodeRedEnv — Emergency Medical Coordination Environment for OpenEnv"
readme = "README.md"
requires-python = ">=3.10"
license = "Apache-2.0"
authors = [{ name = "Darsh" }]
dependencies = [
    "openenv-core>=0.2.1",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "openai>=1.56.0",
    "httpx>=0.27.0",
    "networkx>=3.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
server = "server.app:main"

[tool.hatch.build.targets.wheel]
packages = ["server", "server/models", "server/subsystems"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
